% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsi_functions.R
\name{add_rule}
\alias{add_rule}
\title{add_rule}
\usage{
add_rule(
  fsi,
  antecedents,
  consequents,
  weights = 1,
  logical_op_ant = "AND",
  logical_op_conseq = "AND"
)
}
\arguments{
\item{p}{}
}
\value{

}
\description{

}
\examples{

}
