% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{component_from_sfg}
\alias{component_from_sfg}
\title{Create an empty Plateau Geometry object}
\usage{
component_from_sfg(sfg, md)
}
\arguments{
\item{type}{Type of Plateau Geometry to be created..}
}
\value{
A pgeom class object.
}
\description{

}
\examples{

pgeom_point <- create_empty_pgeom("PLATEAUPOINT")
pgeom_line <- create_empty_pgeom("PLATEAULINE")
pgeom_region <- create_empty_pgeom("PLATEAUREGION")

}
