% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{create_component}
\alias{create_component}
\title{create_component}
\usage{
create_component(raw_obj, md, type)
}
\arguments{
\item{raw_obj}{A vector, list or matrix containing the points to create an sfg object.}

\item{md}{A numeric value indicating the membership degree of the component.}

\item{type}{A character value that indicates the type of the desired sfg object. It can be either \code{"POINT"}, \code{"LINE"}, or \code{"REGION"}.}
}
\value{
A \code{component} object that can be added to a spatial plateau object.
}
\description{
create_component builds a component composed of a Spatial Feature Geometry (sfg) object and its corresponding membership degree.
}
\details{
This function creates an S4 object named \code{component}, which is a pair of an sfg object and a membership degree in \eqn{]0, 1]}.

The spatial data type (i.e., the type of the sfg object) indicated by the parameter \code{type} groups simple and complex objects.
For instance, \code{"POINT"} refers to simple or complex point objects (internally, we can create a POINT or MULTIPOINT object).
}
\examples{

# Creating two components of the type POINT
v1 = rbind(c(1,2), c(3,4))
v2 = rbind(c(1,4), c(2,3),c(4,4))

md1 = 0.2
md2 = 0.1

comp1 <- create_component(v1, md1, type="POINT")
comp2 <- create_component(v2, md2, type="POINT")

# Creating two components of the type LINE

md3 = 0.45
md4 = 0.32

v3 = rbind(c(2,2), c(3,3))
v4 = rbind(c(1,1), c(3,2))

comp3 <- create_component(v3, md3, type="LINE")
comp4 <- create_component(v4, md4, type="LINE")

# Creating two components of the type REGION

p1 <- rbind(c(0,0), c(1,0), c(3,2), c(2,4), c(1,4), c(0,0))
p2 <- rbind(c(1,1), c(1,2), c(2,2), c(1,1))
list_pols_1 <- list(p1,p2)


p3 <- rbind(c(1,0), c(2,0), c(4,2), c(3,4), c(2,4), c(1,0))
p4 <- rbind(c(2,2), c(2,3), c(3,4), c(2,2))
list_pols_2 <- list(p3,p4)

comp_pol1 <- create_component(list_pols_1, 0.4, "REGION")
comp_pol2 <- create_component(list_pols_2, 0.6, "REGION")

}
