% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{create_empty_pgeom}
\alias{create_empty_pgeom}
\title{Creation of an empty \code{pgeom} object}
\usage{
create_empty_pgeom(type)
}
\arguments{
\item{type}{A character value indicating the data type of the \code{pgeom} object.
It can be either \code{"PLATEAUPOINT"}, \code{"PLATEAULINE"} or \code{"PLATEAUREGION"}.}
}
\value{
A \code{pgeom} object.
}
\description{
This function builds an empty \code{pgeom} object of a specific type.
}
\details{
The \code{create_empty_pgeom} creates a new pgeom object with no components. To add new components to this object, you
should use \code{spa_add_component}. The components added to this object must be of same type of the empty pgeom object.
}
\examples{

# Creating an Empty Plateau Point object
empty_plateau_point_pgeom <- create_empty_pgeom("PLATEAUPOINT")

# Creating an Empty Plateau Line object
empty_plateau_line_pgeom <- create_empty_pgeom("PLATEAULINE")

# Creating an Empty Plateau Region object
empty_plateau_region_pgeom <- create_empty_pgeom("PLATEAUREGION")

}
