% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{create_pgeom}
\alias{create_pgeom}
\title{create_pgeom}
\usage{
create_pgeom(components, type)
}
\arguments{
\item{components}{A list, tibble or dataframe of componenets.}

\item{type}{The type of Spatial Plateau Geometry Object to be created.}
}
\value{
An pgeom class object.
}
\description{
Create an object of class pgeom (Plateau Geomtry)
}
\examples{

}
\seealso{
\code{\link{create_component}}
}
