% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{f_symm_diff}
\alias{f_symm_diff}
\title{f_symm_diff}
\usage{
f_symm_diff(x, y)
}
\arguments{
\item{spo}{}
}
\value{

}
\description{

}
\seealso{
Other Fuzzy Difference Operators: 
\code{\link{f_bound_diff}()},
\code{\link{fdiff}()}
}
\concept{Fuzzy Difference Operators}
