% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{fdiff}
\alias{fdiff}
\title{f_diff}
\usage{
fdiff(x, y)
}
\arguments{
\item{spo}{}
}
\value{

}
\description{

}
\seealso{
Other Fuzzy Difference Operators: 
\code{\link{f_bound_diff}()},
\code{\link{f_symm_diff}()}
}
\concept{Fuzzy Difference Operators}
