% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsi_functions.R
\name{fsi_add_fsa}
\alias{fsi_add_fsa}
\title{fsi_add_fsa}
\usage{
fsi_add_fsa(fsi, lvar, tbl)
}
\arguments{
\item{fsi}{The FSI model instantiated with the \code{fsi_create} function.}

\item{lvar}{A character value that represents a linguistic variable of the antecedent.}

\item{tbl}{A tibble with spatial plateau objects annotated with linguistic values of the linguistic variable specified by the above \code{lvar} parameter.}
}
\value{
An FSI model populated with a fuzzy spatial antecedent.
}
\description{
fsi_add_fsa adds a fuzzy spatial antecedent to the FSI model
}
\details{
The fuzzy spatial antecedent added by the \code{fsi_add_fsa} function is composed of a linguistic variable and its corresponding \verb{pgeom objects} annotated by linguistic values.
The format of the \code{tbl} parameter is the same as the output of the function \code{spa_creator}, allowing the user to directly provides plateau region objects as input when designing FSI models.
}
