% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsi_functions.R
\name{new_fsi}
\alias{new_fsi}
\title{new_fsi}
\usage{
new_fsi(
  name,
  fsi_type = "mamdani",
  and_method = "min",
  or_method = "max",
  imp_method = "min",
  agg_method = "max",
  defuzz_method = "centroid"
)
}
\value{

}
\description{

}
\examples{

}
