% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spa_functions.R
\name{pcollection_to_pcomposition}
\alias{pcollection_to_pcomposition}
\title{Converting a spatial plateau collection to a spatial plateau composition}
\usage{
pcollection_to_pcomposition(pcollection)
}
\arguments{
\item{pcollection}{A \code{pcollection} object.}
}
\value{
A spatial plateau composition object.
}
\description{
This function gathers all the fuzzy spatial objects of a \code{PLATEAUCOLLECTION} and
reorganizes them into an equivalent fuzzy spatial composition object.
}
\details{
It gives a fuzzy spatial composition object equivalent to the spatial collection object given as input
aggregating all spatial plateau objects by type.
}
\examples{

library(sf)

pts <- rbind(c(1, 2), c(3, 2))
pcp <- component_from_sfg(st_multipoint(pts), 0.3)

lpts <- rbind(c(1, 1), c(1.2, 1.9), c(2, 1))
lcp <- component_from_sfg(st_linestring(lpts), 1)
pline <- create_pgeometry(list(lcp), "PLATEAULINE", is_valid = TRUE)

rpts1 <- rbind(c(0,0), c(1,0), c(3,2), c(2,4), c(1,4), c(0,0))
rpts2 <- rbind(c(1,1), c(1,2), c(2,2), c(1,1))
rcp <- component_from_sfg(st_polygon(list(rpts1, rpts2)), 0.7)
pregion <- create_pgeometry(list(rcp), "PLATEAUREGION", is_valid = TRUE)

pcomposition <- create_pgeometry(list(pcp, rcp), "PLATEAUCOMPOSITION", is_valid = TRUE)
pcollection <- create_pgeometry(list(pcomposition), "PLATEAUCOLLECTION", is_valid = TRUE)

pcollection_to_pcomposition(pcollection)

}
\references{
\href{https://ieeexplore.ieee.org/document/9177620}{Carniel, A. C.; Schneider, M. Spatial Data Types for Heterogeneously Structured Fuzzy Spatial Collections and Compositions. In Proceedings of the 2020 IEEE International Conference on Fuzzy Systems (FUZZ-IEEE 2020), pp. 1-8, 2020.}
}
