% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsr_base.R
\docType{class}
\name{pgeom-class}
\alias{pgeom-class}
\title{An S4 Class for a Spatial Plateau data type}
\description{
An S4 Class for a Spatial Plateau data type
}
\section{Slots}{

\describe{
\item{\code{component}}{A length-one numeric vector}

\item{\code{supp}}{A sfc object storing the union of all components}

\item{\code{type}}{Plateau Geometry Type}
}}

