% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pgeom_as_tibble}
\alias{pgeom_as_tibble}
\title{Transforms a pgeom object into a tibble object}
\usage{
pgeom_as_tibble(pgeom)
}
\arguments{
\item{pgeom}{pgeom object}
}
\value{
A tibble object with two columns: md and geometry
}
\description{

}
\examples{

}
