% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pgeom_plot}
\alias{pgeom_plot}
\title{pgeom_plot}
\usage{
pgeom_plot(
  pgeom,
  base_poly = NULL,
  low = "white",
  high = "black",
  palette = "Greys",
  m = 10,
  ...
)
}
\arguments{
\item{high}{}
}
\value{

}
\description{

}
\examples{

}
