% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_types.R
\docType{class}
\name{pgeometry-class}
\alias{pgeometry-class}
\title{An S4 Class for representing spatial plateau data types}
\description{
An S4 Class for representing spatial plateau data types
}
\details{
It is a superclass for representing spatial plateau data types. A \code{pgeometry} object stores an \code{sfg} object that represents the union
of all crisp spatial objects of its components (i.e., the support).
}
\section{Slots}{

\describe{
\item{\code{supp}}{An \code{sfg} object that stores the union of all spatial objects of the components of the spatial plateau object.}
}}

\references{
\href{https://ieeexplore.ieee.org/document/8491565}{Carniel, A. C.; Schneider, M. Spatial Plateau Algebra: An Executable Type System for Fuzzy Spatial Data Types. In Proceedings of the 2018 IEEE International Conference on Fuzzy Systems (FUZZ-IEEE 2018), pp. 1-8, 2018.}
}
