% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_types.R
\docType{class}
\name{pregion-class}
\alias{pregion-class}
\title{An S4 Class for representing plateau regions (subclass of \code{pgeometry})}
\description{
An S4 Class for representing plateau regions (subclass of \code{pgeometry})
}
\details{
A \code{pregion} object is composed of a list of \code{component} objects and inherits
the attribute \code{supp} from the class \code{pgeometry} (i.e., the support).
}
\section{Slots}{

\describe{
\item{\code{supp}}{It is inherited from \code{pgeometry}.}

\item{\code{component}}{A list of components.}
}}

\references{
\href{https://ieeexplore.ieee.org/document/8491565}{Carniel, A. C.; Schneider, M. Spatial Plateau Algebra: An Executable Type System for Fuzzy Spatial Data Types. In Proceedings of the 2018 IEEE International Conference on Fuzzy Systems (FUZZ-IEEE 2018), pp. 1-8, 2018.}
}
