% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{search_by_md}
\alias{search_by_md}
\title{search_by_md}
\usage{
search_by_md(components, low, high, m)
}
\arguments{
\item{components}{List with components}

\item{low}{First index of the list to start the binary search}

\item{high}{Last element of the list to end the binary search}

\item{m}{Membership Degree searched}
}
\value{
If found, returns a vector with TRUE and the index position.
If not found, returns a vector with FALSE and the first position given (low)
}
\description{
Searches by a given membership degree in a list of components and their respectives md, returning either TRUE
and the index of the component that has that md or FALSE if not found.

This function use Bynary Search to search by the md.
}
\examples{

md_index <- search_by_md(components= toydata, low = 1, high = 10, md=0.4)



}
