% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spa_functions.R
\name{spa_add_component}
\alias{spa_add_component}
\title{spa_add_component}
\usage{
spa_add_component()
}
\arguments{
\item{component}{}
}
\value{

}
\description{

}
\details{
Implementation of \eqn{\alpha}
}
\examples{

### exemple 1...

### example 2...

}
\references{
Carniel, A. C.; Schneider, M.
Spatial Plateau Algebra: An Executable Type System for Fuzzy Spatial Data Types.
In Proceedings of the 2018 IEEE International Conference on Fuzzy Systems
(FUZZ-IEEE 2018), pp. 1-8, 2018. <https://doi.org/10.1109/FUZZ-IEEE.2018.8491565>
}
\seealso{

}
