% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spa_functions.R
\name{spa_area}
\alias{spa_area}
\title{spa_area}
\usage{
spa_area(pgeom)
}
\arguments{
\item{pgeom}{}
}
\value{

}
\description{

}
\examples{

}
\seealso{
Other Spatial Plateau Metric Operations: 
\code{\link{spa_avg_degree}()},
\code{\link{spa_length}()},
\code{\link{spa_ncomp}()},
\code{\link{spa_perimeter}()}
}
\concept{Spatial Plateau Metric Operations}
