% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_operations.R
\name{spa_boundary}
\alias{spa_boundary}
\title{Capturing the fuzzy boundary of a spatial plateau object}
\usage{
spa_boundary(pgo)
}
\arguments{
\item{pgo}{A \code{pgeometry} object of type \code{PLATEAUPOINT}, \code{PLATEAULINE}, or \code{PLATEAUREGION}.}
}
\value{
A \code{pcomposition} object that represents a fuzzy boundary of \code{pgeometry} object given as input.
}
\description{
This function yields the fuzzy boundary of a plateau region object.
}
\details{
It employs the definition of \emph{fuzzy boundary} of a fuzzy region object in the context of spatial plateau algebra (as defined in the references).
The \emph{fuzzy boundary} of a fuzzy region object \code{A} has a heterogeneous nature since it consists of two parts:
\itemize{
\item a fuzzy line object that corresponds to the boundary of the core of \code{A}.
\item a fuzzy region object that comprises all points of \code{A} with a membership degree greater than 0 and less than 1.
}

This means that the function \code{spa_boundary} can yield these two parts of the fuzzy boundary of a plateau region object (the argument \code{pgeometry}):
the boundary plateau line and the boundary plateau region (i.e., returns a \code{pgeometry} object of the type \code{PLATEAUCOMPOSITION} with these two plateau objects).
}
\examples{
library(tibble)
library(FuzzyR)

set.seed(123)

# some random points to create pgeometry objects by using the function spa_creator
tbl = tibble(x = runif(10, min= 0, max = 30), 
             y = runif(10, min = 0, max = 50), 
             z = runif(10, min = 0, max = 100))

classes <- c("category-1", "category-2")
mf1 <- genmf("trapmf", c(0, 5, 20, 35))
mf2 <- genmf("trimf", c(20, 80, 100))

pregions <- spa_creator(tbl, classes = classes, mfs = c(mf1, mf2))
pregions$pgeometry[[1]]
pregions$pgeometry[[2]]

# capturing and showing the boundary of each pgeometry object previously created
(spa_boundary(pregions$pgeometry[[1]])) 
(spa_boundary(pregions$pgeometry[[2]]))

}
\references{
\itemize{
\item \href{https://ieeexplore.ieee.org/document/7737976}{Carniel, A. C.; Schneider, M. A Conceptual Model of Fuzzy Topological Relationships for Fuzzy Regions. In Proceedings of the 2016 IEEE International Conference on Fuzzy Systems (FUZZ-IEEE 2016), pp. 2271-2278, 2016.}
\item \href{https://ieeexplore.ieee.org/document/8491565}{Carniel, A. C.; Schneider, M. Spatial Plateau Algebra: An Executable Type System for Fuzzy Spatial Data Types. In Proceedings of the 2018 IEEE International Conference on Fuzzy Systems (FUZZ-IEEE 2018), pp. 1-8, 2018.}
}
}
