% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spa_functions.R
\name{spa_difference}
\alias{spa_difference}
\title{spa_difference}
\usage{
spa_difference(pgeom1, pgeom2, dtype = "f_diff")
}
\arguments{
\item{pgeom}{}
}
\value{

}
\description{

}
\examples{

}
\seealso{
Other Spatial Plateau Set Operations: 
\code{\link{spa_intersection}()},
\code{\link{spa_union}()}
}
\concept{Spatial Plateau Set Operations}
