% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spa_functions.R
\name{spa_exact_inside}
\alias{spa_exact_inside}
\title{spa_exact_inside}
\usage{
spa_exact_inside(pgeom1, pgeom2)
}
\arguments{
\item{pgeom}{}
}
\value{

}
\description{

}
\examples{

}
\seealso{
Other Spatial Plateau Topological Relationships: 
\code{\link{spa_boundary_pregion}()},
\code{\link{spa_common_points}()},
\code{\link{spa_contains}()},
\code{\link{spa_contour}()},
\code{\link{spa_core}()},
\code{\link{spa_disjoint}()},
\code{\link{spa_equal}()},
\code{\link{spa_exact_equal}()},
\code{\link{spa_inside}()},
\code{\link{spa_meet}()},
\code{\link{spa_overlap}()},
\code{\link{spa_set_classification}()},
\code{\link{spa_support}()}
}
\concept{Spatial Plateau Topological Relationships}
