% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_geom_set_operations.R
\name{spa_flatten}
\alias{spa_flatten}
\title{Flatten a fuzzy spatial collection}
\usage{
spa_flatten(pcol)
}
\arguments{
\item{pcol}{A \code{pcollection} object.}
}
\value{
A spatial plateau collection object.
}
\description{
This function gathers all the fuzzy spatial objects of a \code{PLATEAUCOLLECTION} and
reorganizes them into a single flattened fuzzy spatial object containing a quadruple
(\code{PLATEAUPOINT}, \code{PLATEAULINE}, \code{PLATEAUREGION}, \code{PLATEAUCOMPOSITION}) that preserves the identity of sub-objects.
}
\details{
It gives a single flattened fuzzy spatial object, aggregating all spatial plateau objects by their types.
In the case of a two-level hierarchy, i.e., there is a \code{PLATEAUCOLLECTION} inside the other,
the function is applied recursively in the lower levels until the quadruple is built.
The t-conorm considered in the aggregation is the \code{max} operator.
}
\examples{
library(sf)

pts <- rbind(c(1, 2), c(3, 2))
pcp <- create_component(st_multipoint(pts), 0.3)

lpts <- rbind(c(1, 1), c(1.2, 1.9), c(2, 1))
lcp <- create_component(st_linestring(lpts), 1)
pline <- create_pgeometry(list(lcp), "PLATEAULINE")

rpts1 <- rbind(c(0,0), c(1,0), c(3,2), c(2,4), c(1,4), c(0,0))
rpts2 <- rbind(c(1,1), c(1,2), c(2,2), c(1,1))
rcp <- create_component(st_polygon(list(rpts1, rpts2)), 0.7)
pregion <- create_pgeometry(list(rcp), "PLATEAUREGION")

pcomposition <- create_pgeometry(list(pcp, rcp), "PLATEAUCOMPOSITION")
pcol1 <- create_pgeometry(list(pcomposition), "PLATEAUCOLLECTION")
pcol2 <- create_pgeometry(list(pline, pregion, pcol1), "PLATEAUCOLLECTION")
pcol2

spa_flatten(pcol2)
}
\references{
\href{https://ieeexplore.ieee.org/document/9177620}{Carniel, A. C.; Schneider, M. Spatial Data Types for Heterogeneously Structured Fuzzy Spatial Collections and Compositions. In Proceedings of the 2020 IEEE International Conference on Fuzzy Systems (FUZZ-IEEE 2020), pp. 1-8, 2020.}
}
