% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{spa_get_type}
\alias{spa_get_type}
\title{Getting the type of spatial plateau object}
\usage{
spa_get_type(pgo)
}
\arguments{
\item{pgo}{A \code{pgeometry} object of any type.}
}
\value{
A type of spatial plateau object.
}
\description{
This function gets the type of a spatial plateau object.
It can be either \code{"PLATEAUPOINT"}, \code{"PLATEAULINE"}, \code{"PLATEAUREGION"}, \code{"PLATEAUCOMPOSITION"}, or \code{"PLATEAUCOLLECTION"}.
}
\details{
The \code{spa_get_type} function gets the type of the spatial plateau object given as input.
For instance, if the \code{pgo} is a object of the class \code{ppoint} (subclass of \code{pgeometry}), the return will be "PLATEAUPOINT".
}
\examples{

library(sf)

pts1 <- rbind(c(1, 2), c(3, 2))
pts2 <- rbind(c(1, 1), c(2, 3), c(2, 1))
pcomp1 <- component_from_sfg(st_multipoint(pts1), 0.4)
pcomp2 <- component_from_sfg(st_multipoint(pts2), 0.3)
ppoint <- create_pgeometry(list(pcomp1, pcomp2), "PLATEAUPOINT")

spa_get_type(ppoint)

}
