% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spa_functions.R
\name{spa_intersection}
\alias{spa_intersection}
\title{spa_intersection}
\usage{
spa_intersection(pgeom1, pgeom2, itype = "min")
}
\arguments{
\item{pgeom}{}
}
\value{

}
\description{

}
\examples{

}
\seealso{
Other Spatial Plateau Set Operations: 
\code{\link{spa_difference}()},
\code{\link{spa_union}()}
}
\concept{Spatial Plateau Set Operations}
