% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spa_functions.R
\name{spa_union}
\alias{spa_union}
\title{spa_union}
\usage{
spa_union(pgeom1, pgeom2, utype = "max")
}
\arguments{
\item{pgeom}{}
}
\value{

}
\description{

}
\examples{

}
\seealso{
Other Spatial Plateau Set Operations: 
\code{\link{spa_difference}()},
\code{\link{spa_intersection}()}
}
\concept{Spatial Plateau Set Operations}
