% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsr_base.R
\docType{class}
\name{component-class}
\alias{component-class}
\title{An S4 Class for representing a component of a spatial plateau object}
\description{
An S4 Class for representing a component of a spatial plateau object
}
\section{Slots}{

\describe{
\item{\code{obj}}{A sf data type}

\item{\code{md}}{A membership degree of the component}
}}

