% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsi_functions.R
\name{fsi_add_cs}
\alias{fsi_add_cs}
\title{fsi_add_cs}
\usage{
fsi_add_cs(fsi, lvar, lvals, mfs, bounds)
}
\arguments{
\item{fsi}{The FSI model instantiated with the \code{fsi_create} function.}

\item{lvar}{A character value that represents a linguistic variable of the consequent.}

\item{lvals}{A character vector that represents linguistic values of the linguistic variable of the consequent.}

\item{mfs}{A character vector of corresponding membership functions.}

\item{bounds}{A numeric vector that represents the lower and upper bounds of the consequent domain.}
}
\value{
An FSI model populated with a consequent.
}
\description{
fsi_add_cs adds the consequent to the FSI model
}
\details{

}
