% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsi_functions.R
\name{fsi_create}
\alias{fsi_create}
\title{fsi_create}
\usage{
fsi_create(
  name,
  and_method = "min",
  or_method = "max",
  imp_method = "min",
  agg_method = "max",
  defuzz_method = "centroid",
  default_conseq = NULL
)
}
\arguments{
\item{default_conseq}{optional. If...MUST be a membership function (something returned by genmf)}
}
\value{

}
\description{

}
\examples{

}
