% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsr_base.R
\docType{class}
\name{pgeom-class}
\alias{pgeom-class}
\title{An S4 Class for representing a spatial plateau object}
\description{
An S4 Class for representing a spatial plateau object
}
\section{Slots}{

\describe{
\item{\code{component}}{A list of components}

\item{\code{supp}}{An sfg object that stores the union of spatial objects of the components of the spatial plateau object}

\item{\code{type}}{The data type of the spatial plateau object}
}}

